fix: move network-cut to a different test
