+++ tests/spread/mongodb/lxd/test_network_cut.py/task.yaml
@@ -1,0 +1,7 @@
+summary: test_network_cut.py
+environment:
+  TEST_MODULE: test_network_cut.py
+execute: |
+  tox run -e integration -- "tests/integration/mongodb/ha/$TEST_MODULE" --substrate lxd --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
